feat: add auth

+++ Cinema-app/app/Http/Controllers/MoviesController.php
@@ -36,11 +36,13 @@ class MoviesController extends Controller
 
     public function create()
     {
+        $this->authorize('create',App\Models\Movie::class);
         return view('movies.create');
     }
 
     public function store()
     {
+        $this->authorize('create',App\Models\Movie::class);
         $data = request()->validate([
             'title' => 'required',
             'description' => '',
@@ -62,11 +64,13 @@ class MoviesController extends Controller
 
     public function edit(Movie $movie)
     {
+        $this->authorize('update',App\Models\Movie::class);
         return view('movies.edit',compact('movie'));
     }
 
     public function update(Movie $movie)
     {
+        $this->authorize('update',App\Models\Movie::class);
         $data = request()->validate([
             'title' => '',
             'description' => '',
@@ -80,6 +84,7 @@ class MoviesController extends Controller
 
     public function destroy(Movie $movie)
     {
+        $this->authorize('delete',App\Models\Movie::class);
         foreach ($movie->seances() as $seances ) {
             $seances->delete();
         }
